added browser server close delay
