+++ src/server/serverManager.ts
@@ -3,8 +3,8 @@ import { Disposable } from '../utils/dispose';
 import { WSServer } from './wsServer';
 import { HttpServer } from './httpServer';
 import { StatusBarNotifier } from './serverUtils/statusBarNotifier';
-import { AutoRefreshPreview, GetConfig, UpdateSettings } from '../utils/utils';
-import { DONT_SHOW_AGAIN, Settings } from '../utils/constants';
+import { AutoRefreshPreview, GetConfig, UpdateSettings, Settings } from '../utils/settingsUtil';
+import { DONT_SHOW_AGAIN } from '../utils/constants';
 import { serverMsg } from '../manager';
 
 export interface PortInfo {
