+++ src/utils/constants.ts
@@ -15,19 +15,3 @@ export const OPEN_EXTERNALLY: vscode.MessageItem = {
 	title: 'Open Externally',
 };
 export const HOST = '127.0.0.1';
-
-export const SETTINGS_SECTION_ID = 'LiveServer';
-
-export const Settings: any = {
-	portNum: 'portNum',
-	showStatusBarItem: 'showStatusBarItem',
-	showServerStatusPopUps: 'showServerStatusPopUps',
-	autoRefreshPreview: 'autoRefreshPreview',
-	browserPreviewLaunchServerLogging: 'browserPreviewLaunchServerLogging',
-	openPreviewTarget: 'openPreviewTarget',
-};
-
-export const PreviewType: any = {
-	internalPreview: 'internalPreview',
-	externalPreview: 'externalPreview',
-};
