+++ src/utils/utils.ts
@@ -1,29 +1,4 @@
 import * as vscode from 'vscode';
-import {
-	GO_TO_SETTINGS,
-	Settings,
-	SETTINGS_SECTION_ID,
-	PreviewType,
-} from './constants';
-interface LiveServerConfigItem {
-	portNum: number;
-	showStatusBarItem: boolean;
-	showServerStatusPopUps: boolean;
-	autoRefreshPreview: AutoRefreshPreview;
-	browserPreviewLaunchServerLogging: boolean;
-	openPreviewTarget: OpenPreviewTarget;
-}
-
-export enum AutoRefreshPreview {
-	onAnyChange = 'On All Changes in Editor',
-	onSave = 'On Changes to Saved Files',
-	never = 'Never',
-}
-
-export enum OpenPreviewTarget {
-	embeddedPreview = 'Embedded Preview',
-	externalBrowser = 'External Browser',
-}
 
 export function FormatDateTime(date: Date, delimeter = ', '): string {
 	const mm = date.getMonth() + 1;
@@ -52,33 +27,6 @@ export function FormatFileSize(bytes: number) {
 	return `${modifiedSize} TB`;
 }
 
-export function GetConfig(resource: vscode.Uri): LiveServerConfigItem {
-	const config = vscode.workspace.getConfiguration(
-		SETTINGS_SECTION_ID,
-		resource
-	);
-	return {
-		portNum: config.get<number>('portNum', 3000),
-		showStatusBarItem: config.get<boolean>('showStatusBarItem', true),
-		showServerStatusPopUps: config.get<boolean>(
-			Settings.showServerStatusPopUps,
-			false
-		),
-		autoRefreshPreview: config.get<AutoRefreshPreview>(
-			Settings.autoRefreshPreview,
-			AutoRefreshPreview.onAnyChange
-		),
-		browserPreviewLaunchServerLogging: config.get<boolean>(
-			Settings.browserPreviewLaunchServerLogging,
-			true
-		),
-		openPreviewTarget: config.get<OpenPreviewTarget>(
-			Settings.openPreviewTarget,
-			OpenPreviewTarget.embeddedPreview
-		),
-	};
-}
-
 export function GetRelativeActiveFile(): string {
 	const activeFile = vscode.window.activeTextEditor?.document.fileName;
 	return activeFile ? GetRelativeFile(activeFile) : '';
@@ -90,40 +38,6 @@ export function GetRelativeFile(file: string): string {
 	return ret;
 }
 
-export function SettingsSavedMessage(): void {
-	vscode.window
-		.showInformationMessage(
-			'Your selection has been saved in settings.',
-			GO_TO_SETTINGS
-		)
-		.then((selection: vscode.MessageItem | undefined) => {
-			if (selection === GO_TO_SETTINGS) {
-				vscode.commands.executeCommand(
-					'workbench.action.openSettings',
-					SETTINGS_SECTION_ID
-				);
-			}
-		});
-}
 
-export function UpdateSettings<T>(
-	settingSuffix: string,
-	value: T,
-	isGlobal = true
-): void {
-	vscode.workspace
-		.getConfiguration(SETTINGS_SECTION_ID)
-		.update(settingSuffix, value, isGlobal);
-	SettingsSavedMessage();
-}
 
-export function GetPreviewType(extensionUri: vscode.Uri): string {
-	if (
-		GetConfig(extensionUri).openPreviewTarget ==
-		OpenPreviewTarget.embeddedPreview
-	) {
-		return PreviewType.internalPreview;
-	} else {
-		return PreviewType.externalPreview;
-	}
-}
+
